1853 - Reverted part of PR 1825 which moved `split_s3_uri`
