+++ projects/_01_ingest/ascwds/fargate/Dockerfile
@@ -11,7 +11,6 @@ COPY polars_utils polars_utils
 COPY projects/_01_ingest/ascwds/fargate/utils projects/_01_ingest/ascwds/fargate/utils
 COPY utils/column_names utils/column_names
 COPY utils/column_values utils/column_values
-COPY utils/s3_file_utils.py utils/s3_file_utils.py
 
 # Copy all python jobs to WORKDIR
 COPY projects/_01_ingest/ascwds/fargate/*.py .
