+++ tests/unit/test_utils.py
@@ -431,6 +431,13 @@ class GeneralUtilsTests(UtilsTests):
         returned_data = returned_df.collect()
         self.assertEqual(expected_data, returned_data)
 
+    def test_split_s3_uri(self):
+        s3_uri = "s3://sfc-data-engineering-raw/domain=ASCWDS/dataset=workplace/"
+        bucket_name, prefix = utils.split_s3_uri(s3_uri)
+
+        self.assertEqual(bucket_name, "sfc-data-engineering-raw")
+        self.assertEqual(prefix, "domain=ASCWDS/dataset=workplace/")
+
     def test_create_unix_timestamp_variable_from_date_column(self):
         column_schema = StructType(
             [
