--- utils/s3_file_utils.py
@@ -1,12 +1,0 @@
-def split_s3_uri(uri: str) -> tuple[str, str]:
-    """
-    Converts a given string of an s3 uri into its bucket and key names
-
-    Args:
-        uri (str): The s3 uri to be split.
-
-    Returns:
-        tuple[str, str]: A tuple of the bucket and key substrings from the s3 uri.
-    """
-    bucket, prefix = uri.replace("s3://", "").split("/", 1)
-    return bucket, prefix
